Remove old files and reorganize
