+++ src/Application/Models/Asset.js
